requirements.txt
asgiref==3.5.0
Django==3.1.4
psycopg2==2.9.3
sqlparse==0.4.2
Pillow==9.1.0
djangorestframework==3.13.1
dropbox==11.30.0
pytz==2022.1
six==1.16.0
gunicorn==20.1.0
dj-database-url==0.5.0
boto3==1.22.3
django-storages==1.12.3
whitenoise==6.0.0
requests==2.27.1
pandas==1.4.2
XlsxWriter==3.0.3
django-allauth==0.50.0
social-auth-app-django==5.0.0
